Lint files, remove dead-code, slightly rework folder structure.
  Nesting the server logic under a namespace 'server' makes for nicer
  imports and tidy a bit the overall project structure instead of having
  everything in one folder.
