+++ tests/health_check.rs
@@ -1,6 +1,6 @@
 //! tests/health_check.rs
 use std::net::TcpListener;
-use tokenapi;
+use tokenapi::server;
 
 // spawn_app runs the application in the background so we can run tests
 // against it. Should that server fail to create there is no need to
@@ -14,11 +14,11 @@ fn spawn_app() -> String {
     let ip = listener.local_addr().unwrap().ip();
     let mappings = String::from("./registry/mappings");
 
-    let server = tokenapi::run(listener, mappings).expect("Failed to create server");
+    let server = server::run(listener, mappings).expect("Failed to create server");
 
     // tokio spawn takes a future and hands it over to its runtime for
     // continious polling
-    let _ = tokio::spawn(server);
+    tokio::spawn(server);
     format!("http://{}:{}", ip, port)
 }
 
@@ -77,38 +77,16 @@ async fn subject_endpoint_returns_200() {
 
     assert!(response.status().is_success());
     assert_eq!(response.status().as_u16(), 200);
-    // let body = response.text().await.unwrap();
-    //let p = PathBuf::from("../registry_data/782c158a98aed3aa676d9c85117525dcf3acc5506a30a8d87369fbcb4d6f6e6574.json");
-    //let mut json_file = File::open(p)?;
-    //let mut json = String::new();
-    //json_file.read_to_string(&mut json).unwrap();
-    //assert_eq!(json, body);
 }
 
 #[tokio::test]
 async fn query_endpoints_returns_400() {
-    struct QueryPayload {
-        subjects: Vec<String>,
-        properties: Option<Vec<String>>,
-    }
     let address = spawn_app();
 
     let client = reqwest::Client::new();
-    let payload = QueryPayload {
-        subjects: vec![
-            String::from("986f0548a2fd9758bc2a38d698041debe89568749e20ab9b75a7f4b7444149"),
-            String::from("d6a8d8af07d704ba941aa1e4095cbb6968e45ccd3e70340867bf9b083138315261726f417a756c47656c6f")
-        ],
-        properties: Some(vec![
-            String::from("subject"),
-            String::from("name"),
-            String::from("description")
-        ])
-    };
 
     let response = client
         .post(&format!("{}/metadata/query", &address))
-        // .json(&payload)
         .send()
         .await
         .expect("Dies!");
